ensure should refresh the certificate if it removes them
Currently when the spec is updated, ensure removes the certificate, but because
the provider is loaded before that with a valid certificate still,
`ensure` will not reissue the cert so we have missing certs and have to
run `ensure` again to reissue the cert.

This patch changes it so that whenever we remove any certificate,
`spec.Ready` will return false so we can add the cert into the queue.