cert: add support for writing CA certificates to disk.
An oft-requested feature is to allow certmgr to write the CA certificate
to disk. This change provides support for that. If no file section is
provided in the CA configuration (as is the case in all current
configurations), certmgr's behaviour is unmodified with the exception of
a log message noting that the certificate will not be written.