Add a simple/stupid 'command' service manager.
If a cert specifies 'service_manager' setting as 'command', then action
is passed to /bin/bash -c (or /bin/sh if bash can't be found).  This allows
custom actions to be taken- things beyond just init scripts.

Note that there are multiple CERTMGR_* variables exposed to the shell code
invoked; this is to allow for the target to decide what to do when these
things change.