Fix certmgr refreshing logic for when the spec has been updated.
This sequence would fail and result in they key existing, but the cert missing.
Another `certmgr ensure` invocation would be required to repair it.

* touch my-definition
* certmgr ensure

The cause is that certmgr would try and force cfssl to update the cert/key
via unlink'ing the cert.  The problem here is that cfssl maintains in memory
timestamps to determine when something should be renewed- and as far as it
was concerned, the cert was still valid, thus wouldn't regenerate it.  The underlying
provider layer also wouldn't force a refresh since it thinks the content is still on
disk.

Thus this approach; reach in and reset timestamps in cfssl so that when it's asked to
refresh, it thinks a refresh is due- and actually does the refresh.