Fix certmgr ensure to only trigger actions when content has changed.
First, CA.loaded was removed because it wasn't used for anything.

Second, CA.Load() was gutted so it no longer writes the CA to disk- that's
CA.Refresh()'s job, and it's *critical* that Refresh() handle this since
manager instances need to know if actions should be triggered.

Third, strip space from the CA when we read it from disk- we seem to be picking up
a trailing newline from WriteFile (which is good for humans) but it breaks our
byte comparison for what we receive from the remote.

The test case to trigger this basically comes down to just having an action in place,
and invoking `certmgr ensure --debug` twice; the first run may update things as needed,
but the second run *should not trigger* an action, but would.

With these logic changes, this no longer happens.