Add `certmgr ensure --forceRegen` to override lifespan checks, and force update.
This allows you to pre-emptively force a regen, without having to wipe
the materials from disk first.  This is useful for certain operational
steps.  It shouldn't default on, but it should be possible to override
the internal calculation without removing PKI material from disk to force
the run.

From a development and debugging standpoint, this is also useful for
when you're iterating on a spec's definition (action in particular) and
need to force certmgr to run so you can validate it.

Again, one can accomplish the same via wiping the cert/key from disk,
but that's a crappy workflow and it has some operational risks that
may not be desirable- for example, if the process has an inotify in
place for tracking the key/cert.