Fix invalid action triggering when CA has changed but we don't write it to disk.
In 1f1c919d083d12d4fcf, multiple CA bugs were fixed- mostly issues where
certmgr was triggering an action for a spec because it thought the CA had changed
(but it had not).

That commit moved the updating of the internal notion of the CA into
the file serialization method; the problem is that that method was only
invoked if the CA was being serialized to disk.

If it wasn't being serialized to disk, if the CA changed it would result
in certmgr triggering the action every time it checked that particular spec.

Without this commit, the only work around is to restart certmgr itself.

This bug affected v1.6.1, v1.6.2, and v1.6.3.