Tolerate starting without any specs defined.
Previously, certmgr daemon mode would refuse to run if it couldn't find
any specs to manage; you literally couldn't start the daemon in that
case.

While running it without specs is a bit daft, if one deploys certmgr
as part of a standard fleet image- as an expected service- that behaviour
becomes problematic if the host doesn't (yet) have any specs configured.

Thus this change: it converts "no specs found" into a warning, but
tolerates it for daemon mode.

If folks want the old behaviour, they can pass --requireSpecs to the
daemon invocation and it'll behave as it did before.

Finally, `certmgr ensure` also got a tweak; if you're invoking ensure
directly, you're doing this because there should be specs you want
processed.  That behaviour makes sense, but is now configurable.

If you invoke `certmgr ensure` and it fails to find any specs to process,
that results in a non-zero exit code.

If you want to tolerate no specs being found, you can invoke
`certmgr ensure --allowZeroSpecs` to allow this.

A warning is logged either way, but it'll not fail if no specs were
found and that flag was enabled.