Incorporate the expiry time of a cert/CA directly as a metric.
In the process, transfer the 'age' labels out of SpecWatchCount- prometheus
can work with timestamps as a metric value, but can't work with it as a label,
so they're not as useful as a raw metric.

Using a raw metric like this in turn allows us to track exactly when a
spec has it's next expire even- and to track and alert if certmgr hasn't
done anything about it.

Finally, via expanding the expires metric- making it per spec- this allows
removal of the SetExpireNext method which presented a summarized version
of the metric; it's better to expose the full details and let folks doing
prometheus query do min() to get the next expiry.

Sans that, also wipe yet more dead methods from cert related to this code.