Cleanup Manager instantiation, require NewFromConfig format to be yaml.
Previous code tried to support json format for certmgr.yaml- it's
not worth it, just require it to be yaml.

In the process, I'm cleaning up the shadowing of interval/before
that folks wrote to work around this.

In the future for things like this, just add an UnmarshalYAML and
properly parse it.