Cleanup spec reload logic.
Due to the intertwined nature, this is one change that is essentially
two.

From the spec side, the mtime comparisons for key/cert against spec have
been removed- x509 comparisons are a far better approach, and reinforcing those
pathways is what we should invest in.

Unfortunately, cfssl still will not regenerate a key that has changed algo/keysize,
but that's a pre-existing issue- and one that will be fixed in follow up PRs.

These changes require shifting more internals of Spec to actual internal logic,
which leads to the manager changes.

The manager now asks the spec if there are changes on disk; if there are changes,
the manager in turn creates a new spec instance.  If that instance parses, we
use that- replacing the old.  If that spec doesn't parse, we use the old instance
already in memory.

This ensures that we don't corrupt the instance in memory from a failed loading of
a changed spec on disk.

Finally, the manager changes codify a subtle fact that certmgr doesn't properly
watch the directory of spec's, thus was never able to detect new content beyond
by chance.  This is fixable via integration of fsnotify, but that's something
for a later PR- in the interim, this watches just the spec's it saw at startup
and does reload of just those.

Per the comments, there is followup work; this commit doesn't make existing
behaviour any worse than it was already (for what's already bad), but does
fix behaviour where we can- and positions the code so those issues can be
more easily addressed.