Add check that the cert we're working with is still signed by the CA.
If the CA has changed, under the certmgr model the cert should be regenerated.

In a perfect world, this would only regenerate if the cert was no longer
covered by the CA bundle- old and new CA- but that's not the model for
cfssl, thus this approach.

Most folks just update the CA while keeping the serial- thus this check
won't regenerate a cert if that's the CA rotation model.

If however they're changing the serial, this catches that the cert
is no longer covered and will trigger a regeneration.