Check for OOB permission changes to PKI on disk every run.
If the permissions has changed, the consuming service may not be able
to access it- thus trigger a regeneration in that case so the permissions
get properly updated.

Note that this *can* lead to a war of certmgr trying to enforce permissions
it has been told to enforce, and an external SCM trying to change the permissions.

This is inherintly racy and wrong either way- and certmgr has no way to know
if the service actually was able to use the PKI- thus the only option is
to just force a regen.