Run spec's on their own timer, and support differing Before/Interval for each.
The original code of certmgr was a central 'wake up, run the specs all at once'-
this is a model that hasn't worked, and is one workloads have long since grown past.

Spec's need to be able to configure their own variant Before Duration- a 30 day TTL
doesn't need 3 weeks of Before, but a 9 month one- yeah, it makes sense.

The same applies for Interval- it should be per spec.

To support this- and future additions like splay logic- the manager is again
being gutted of core logic and moved towards an aggregator of spec's.  Specifically,
Spec's now have their own Run method that does waiting and enforcing- the manager just
starts those and waits for their exit.

To support this, manager defaults are now managed via a SpecOptions struct; in doing so,
this also moves the various spec options (which I'll be adding to) out of function protoypes
and into struct's.  We break API compatability, but we replace it with a mechanism
that won't break in the future when we extend options.

Final note: with this shift of the 'focus'- this fix of the abstractions- there is
yet more code that should be ripped out or made private.  That's follow up work.