Make configuration reloads deterministic via SIGHUP support.
Previous logic would do scans of the spec files (but not manager files)
during the interval wakeup.  It was only able to act upon changed specs-
not new spec's, and it couldn't handle deleted spec's properly.  That was
a known fault, but was an intermediate step moving away from the original
form to what is now in place.

With this change, configuration reload is purely deterministic- it won't
happen unless you tell it to do so, and it only will 'activate' that new
manager/spec config's if it can be properly/successfully reloaded.

This is triggered via a SIGHUP.

As a consequence of this, the code will still check spec's for size/mtime
difference and will log warnings if config is out of sync with what's in
memory- longer term, this will likely be removed.

Finally, proper signal handling for shutdown was also added- this is to
ensure that certmgr actually does a graceful shutdown rather than just
dieing in the middle of writing PKI to disk (for example).