Provide a configurable to prevent certmgr from restarting downed services.
If PKI expires and an action is configurable, certmgr will fire that action.

If the action is 'restart'- it'll lead to an intentionally downed service
being started, which is less than optimal.

To enable this new behaviour, set `take_actions_only_if_running: true` in
the spec (or manager config).