Slight refactor of configuration parsing
The intermediate object "manager" was removed, as it was unused.
The intermiate variables for logging have been removed, and replaced with calls to
look up the setting in viper. This fixes configuring logging from config files

The configuration option and cmdline parameter interval.splay has been renamed intervalsplay
This fixes an issue where the 'interval' config would not work if you also specified 'interval: splay:'

The configuration dumping in the 'version' output has also been updated to show the values of the splay settings.